requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9          # or psycopg[binary] if using psycopg 3
python-dotenv==1.0.1            # optional — for loading .env
gunicorn==22.0.0